requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
python-pptx==0.6.23
google-generativeai==0.3.2 